Crate evolution

An Evolution (game) competition library.

Re-exports

pub use self::game::Game;

Modules

ext
Extension methods provided by traits for built-in types.

game
Types and functions required for playing a game of Evolution.

interact
Data and methods for required interactions for the game.

object
Objects of the game, like a card or a species.

silly
A very silly chooser.

Macros

debug_expr
A helper that compiles into an evaluation of an expression, but also logs the expression and evaluated result to the debug logs before returning.

test_harness
A helper that compiles into a test harness for portions of the wire data types.