Struct evolution::game::context::Board

pub struct Board<'a, C: 'a + Chooser>(_);

A game aware board.

Methods

impl<'a, C: Chooser> Board<'a, C>

pub fn new(game: &'a mut Game<C>) -> Self

Create a new contextual board instance from a mutable Game object reference.

Arguments

game - A mutable reference to a Game instance.

Returns

A contextual Board instance referencing the given Game instance.

pub fn reveal(&mut self)

Reveal the cards played to the board as food. Doing so will also clear the cards from the board and add or subtract the appropriate number of FoodTokens.

Arguments

&mut self - The contextual board instance to reveal cards for.

Panics

This function panics if there are no cards to reveal.