Struct evolution::game::context::Species

pub struct Species<'a, C: 'a + Chooser>(_, _, _);

A game aware species.

Methods

impl<'a, C: Chooser> Species<'a, C>

pub fn new(game: &'a mut Game<C>, indices: (usize, usize)) -> Self

pub fn player(&mut self) -> Player<C>

This species' player.

pub fn neighbor(&mut self, placement: Placement) -> Option<Species<C>>

This species' neighbor on the given placement.

pub fn feed(&mut self)

Feed this species.

pub fn eat(&mut self) -> bool

Tell this species to eat, returning false if they can't.

pub fn cooperate(&mut self)

Trigger cooperation for this species if applicable.

pub fn store(&mut self, amount: u64)

Store the given amount of food into this species' fat.

pub fn kill(&mut self) -> Result<bool, ()>

Kill a member of this species.