Struct evolution::game::step::Bag

pub struct Bag<'a, C: 'a + Chooser>(pub &'a mut Game<C>);

cull all of the species in the game, giving 2 cards to each player per species that went extinct. Then bag the food for all players.