Module evolution::interact

Data and methods for required interactions for the game.

An interaction is when a part of the system needs to share some information and possibly get a response. For example the dealer will ask a player for it's choice for feeding each turn.

Examples

Making a choice for feeding.

use evolution::Game;
use evolution::object::{FoodToken, Placement};
use evolution::silly::Silly;
use evolution::interact::{Observe, Choose, FeedChoice};

let mut game = Game::<Silly>::new(3).unwrap();
game.board_mut().push_food(FoodToken);
game.players_mut()[0].domain_mut().add(Placement::Right);

if let Ok(Some(choice)) = Silly.choose(&game.observe()) {
    assert_eq!(FeedChoice::Feed(0), choice);
} else {
    panic!("must have gotten a choice");
}

Re-exports

pub use self::observe::Observe;

pub use self::observe::Observation;

pub use self::choose::Choose;

pub use self::choose::Choice;

pub use self::choices::Choices;

Modules

choices
Public to get around https://github.com/rust-lang/rust/pull/31920.

choose
Public to get around https://github.com/rust-lang/rust/pull/31920.

observe
Public to get around https://github.com/rust-lang/rust/pull/31920.

Structs

ActionChoice
A choice how to use cards for a round.

ActionObservation

BoardObservation
The observed view of a board, with just the food of a board.

BoardTrade
A choice to trade a card for a species board, and optionally apply traits to it.

DealObservation

FeedObservation

Growth
A choice to grow a species in either body size or population by trading a card.

PlayerObservation
The visable state of a player.

TraitTrade
A choice to swap out a trait on a species for a new one.

Enums

FeedChoice
A choice of a species to feed.

Traits

Chooser