Trait evolution::interact::observe::Observe

pub trait Observe<O: Observation> {
    fn observe(&self) -> O;
}

Asking some object for an observation of itself.

This is used, for example to get an observation of the game for a player to choose a response for feeding.

Required methods

fn observe(&self) -> O

Return the observation of this data.

Loading content...