Module evolution::object

Objects of the game, like a card or a species.

These objects contain complete knowledge, this means that they represent the actual object in the game, not for example just the public information.

Structs

Board
An owned board with food tokens, and cards which were played as food for the round.

Card
A card in evolution.

Domain
A collection of species for a player.

FoodToken
A structure enforcing single ownership over all food tokens in the game world.

Player
The internal state of a player, including private information.

Species
A species which evolves, feeds, grows, and fluctuates in population throughout the game.

Enums

Placement
A placement describes where players can place a new species in relation to their other species.

Trait
A trait which a trait card may have. These are modifiers for how a species board acts.

Constants

CARDS_PER_EXTINCTION
The number of cards awarded to a player for each extinct species.

HARD_SHELL_PROTECTION
The effective addeded body size the hard shell trait adds, for deflecting attacks.

MAX_BODY_SIZE
The maximum body size a species can have.

MAX_POPULATION
The maximum population a species can have.

MAX_TRAITS
The maximum number of traits a species may have.

NUM_CARNIVORE_CARDS
The number of carnivore trait cards in a deck.

NUM_VEGITARIAN_CARDS
The number of each non-carnivore cards in a deck.