Struct evolution::object::Board

pub struct Board { /* fields omitted */ }

An owned board with food tokens, and cards which were played as food for the round.

Methods

impl Board

Food functions.

pub fn food(&self) -> &[FoodToken]

Returns the amount of food available on the board. This is the watering hole value as described in the spec.

Arguments

&self - The board to get food from.

Returns

An immutable slice of type &[FoodToken], containing the food tokens on the board currently.

pub fn push_food(&mut self, food_token: FoodToken)

Move a token onto the board.

Arguments

&mut self - The board taking food.

food_token - A food token for the board to take.

pub fn push_foods(&mut self, food_tokens: Vec<FoodToken>)

Move many tokens onto the board.

Arguments

&mut self - The board taking food.

food_tokens - A vector of food tokens for the board to take.

pub fn pop_food(&mut self) -> Option<FoodToken>

Get a single food token, if available. An empty watering hole will return None.

Arguments

&mut self - The board to giving food.

Return

This function returns a food token from the board as Some or None if there are no more food tokens left.

pub fn pop_foods(&mut self, requested: u64) -> Option<Vec<FoodToken>>

Removes many food tokens from the board. This function returns None if there are no food tokens to take. If this function is asked for more food than the board has this function will return as much as it can.

impl Board

Card functions

pub fn cards(&self) -> Option<&[Card]>

Get the cards used as food for this round.

Arguments

&self - The board we are asking for the cards of.

Return

This function returns None before the cards are flipped. Cards are flipped after the play phase but before the eat phase, when flipped then this function returns a slice of the cards.

pub fn cards_mut(&mut self) -> Option<&mut [Card]>

Get the cards used as food for this round.

Arguments

&mut self - The board we are asking for the cards of.

Return

This function returns None before the cards are flipped. Cards are flipped after the play phase but before the eat phase, when flipped then this function returns a slice of the cards.

pub fn add_card(&mut self, card: Card)

Adds a single card to the board to be used as food.

Arguments

&mut self - The board to add a card to.

card - The card to add to the board.

pub fn set_cards(&mut self, cards: Vec<Card>)

Set the cards used as food for this round.

Arguments

&mut self - The board to set cards for.

Panics

This function will panic if there are already cards on the board.

pub fn clear_cards(&mut self)

Clears the cards used as food for this round.

Arguments

&mut self - The board to clear cards off of.