Struct evolution::object::Species

pub struct Species { /* fields omitted */ }

A species which evolves, feeds, grows, and fluctuates in population throughout the game.

Population must be from 0 to MAX_POPULATION, and body size must be from 0 to MAX_BODY_SIZE. Traits must be a collection of 0 to MAX_TRAITS length, and each member of the collection must be unique. Food must not be larger than the population, and fat is only Some if the species has FatTissue trait, and must not be greater than the body size.

A species "eating" means that it is consuming one food token. This is distinct from "feeding" which is a process of eating in which the species and other spcies may eat.

A species "storing" means that the species is consuming some number of food tokens and putting them on their fat tissue. This is only allowed if the species has the trait.

Examples

The default species.

use evolution::object::*;

let species = Species::default();
assert_eq!(1, species.population());
assert_eq!(0, species.body_size());
assert!(species.traits().is_empty());
assert!(species.food().is_empty());
// TODO: `assert!(species.fat().is_empty());` really shouldn't panic.

Methods

impl Species

Population functions.

pub fn population(&self) -> u64

Returns the current population of this species. If this value is 0 then this species is considered extinct.

Examples

use evolution::object::*;

let spec = Species::default();
assert_eq!(1, spec.population());

pub fn is_extinct(&self) -> bool

Returns true if this species is extinct.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.kill().unwrap();
assert!(spec.is_extinct());

pub fn breed(&mut self) -> Result<u64, ()>

Increase the population of this species by one. This function returns an Err if the population is already the maximum value.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.breed().expect("unable to breed");
assert_eq!(2, spec.population());

pub fn kill(&mut self) -> Result<bool, ()>

Kill one this species, removing 1 food from the species if needed to maintain the invariant that a species must have food <= population.

This function assumes it will never be called on an extinct species. In general it's best to never allow data structures to contain species with a population of 0.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.kill().unwrap();
assert_eq!(0, spec.population());

pub fn cull(&mut self) -> Result<bool, ()>

Kills of as many of this species as needed to make the population equal to the the food. Returns true if this species went extinct.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.breed().unwrap();
spec.breed().unwrap();
spec.eat(FoodToken).unwrap();
assert_eq!(3, spec.population());
spec.cull().unwrap();
assert_eq!(1, spec.population());

impl Species

Body size functions.

pub fn body_size(&self) -> u64

Returns the body size for this species.

Examples

use evolution::object::*;

let mut spec = Species::default();
assert_eq!(0, spec.body_size());
spec.grow().expect("unable to grow");
spec.grow().expect("unable to grow");
assert_eq!(2, spec.body_size());

pub fn grow(&mut self) -> Result<u64, ()>

Grow a species, increasing their body size by one. This function returns Err if the body size is already the maximum value.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.grow().expect("unable to grow");
assert_eq!(1, spec.body_size());

impl Species

Trait functions.

pub fn traits(&self) -> &[Trait]

Returns the traits in the order they were added.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.evolve(Trait::Climbing).expect("unable to evolve");
assert_eq!(1, spec.traits().len());

pub fn has_trait(&self, t: Trait) -> bool

Returns true if this species has the given trait.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.evolve(Trait::Burrowing).expect("unable to evolve");
assert!(spec.has_trait(Trait::Burrowing));
assert!(!spec.has_trait(Trait::Climbing));

pub fn evolve(&mut self, t: Trait) -> Result<(), ()>

Give this species the given trait. This function returns Err when this species already has the requested trait.

Examples

use evolution::object::*;

let mut spec = Species::default();
assert_eq!(0, spec.traits().len());
spec.evolve(Trait::Climbing).expect("unable to evolve");
spec.evolve(Trait::FatTissue).expect("unable to evolve");
assert_eq!(2, spec.traits().len());
assert!(spec.has_trait(Trait::Climbing));
assert!(spec.has_trait(Trait::FatTissue));

pub fn exchange_trait(&mut self, i: usize, t: Trait) -> Result<(), ()>

Swaps a trait in-place for this species. Returns an error if the given index is out-of-bounds.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.evolve(Trait::Foraging).expect("unable to evolve");
assert!(spec.has_trait(Trait::Foraging));
assert!(spec.exchange_trait(0, Trait::Carnivore).is_ok());
assert!(spec.has_trait(Trait::Carnivore));

impl Species

Food functions.

pub fn food(&self) -> &[FoodToken]

Returns the amount of food this species currently has.

Examples

use evolution::object::*;

let mut spec = Species::default();
assert_eq!(0, spec.food().len());
spec.eat(FoodToken).expect("unable to eat");
assert_eq!(1, spec.food().len());

pub fn can_eat(&self) -> bool

Returns true if this species can eat one or more food tokens.

Examples

use evolution::object::*;

let mut spec = Species::default();
assert!(spec.can_eat());
spec.eat(FoodToken).expect("unable to eat");
assert!(!spec.can_eat());

pub fn eats(&self) -> usize

Returns the number of eats a player's species should perform for a feed.

pub fn eat(&mut self, food_token: FoodToken) -> Result<(), FoodToken>

Directs this species to eat a single food token. If the species was unable to eat the food token, the food token is returned in Err, otherwise Ok(()) is returned.

Examples

use evolution::object::*;

let mut spec = Species::default();
assert_eq!(0, spec.food().len());
spec.eat(FoodToken).unwrap();
assert_eq!(1, spec.food().len());

pub fn take_food(&mut self) -> Vec<FoodToken>

Return all the food from this species.

impl Species

Fat functions.

pub fn fat(&self) -> &[FoodToken]

Returns the amount of fat food stored.

Examples

Calling fat() on a species with fat tissue.

use evolution::object::*;

let mut species = Species::default();
assert_eq!(0, species.fat().len());
species.evolve(Trait::FatTissue).unwrap();
species.grow().unwrap();
species.store(vec![FoodToken]).unwrap();
assert_eq!(1, species.fat().len());

pub fn can_store(&self) -> Option<u64>

Returns true if this species can store the given number of food tokens.

Examples

// In debugging this test, trying to find the line number I noticed the
// line number reported by `cargo test` was 2 larger than the actual
// line number for the test failure. I'm guessing it's for something
// like this.
//
// #[test]
// fn can_store_0() {
use evolution::object::*;

let mut spec = Species::default();
assert_eq!(None, spec.can_store());
spec.evolve(Trait::FatTissue).unwrap();
spec.grow().unwrap();
assert_eq!(Some(1), spec.can_store());
assert!(spec.store(vec![FoodToken]).is_ok());
// }

pub fn store(&mut self, food_tokens: Vec<FoodToken>) -> Result<(), ()>

Store the given food tokens to this species fat. This function returns an error if the species cannot store the food, either because it doesn't have the trait, or because too much food was given.

pub fn digest_fat(&mut self)

Move food from fat to food. This function can only be called on species with the FatTissue trait.

Examples

use evolution::object::*;

let mut spec = Species::default();
spec.evolve(Trait::FatTissue).unwrap();
spec.grow().unwrap();
spec.store(vec![FoodToken]).unwrap();

assert_eq!(0, spec.food().len());
assert_eq!(1, spec.fat().len());
spec.digest_fat();
assert_eq!(1, spec.food().len());
assert_eq!(0, spec.fat().len());

impl Species

Attacking functions.

pub fn can_attack(
&self,
target: &Species,
left: Option<&Species>,
right: Option<&Species>
) -> bool

Returns true if this species can attack the given target species, given it's neighbors. The traits of each card dictate the result.

Undefined Behavior

The behavior of this function is not defined for attacking extinct target species.

Examples

use evolution::object::*;

let mut attacker = Species::default();
attacker.evolve(Trait::Carnivore).expect("unable to evolve");
let defender = Species::default();
// attacker can attack defender with no neighbors to help.
assert!(attacker.can_attack(&defender, None, None));